Move start/stop test db to suite-level
